Move last function out of constants
